+++ packages/terra-aggregator/src/terra-dev-site/doc/example/CombinedDisclosureExample.jsx
@@ -1,0 +1,41 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import ModalManager from 'terra-modal-manager';
+import SlidePanelManager from 'terra-slide-panel-manager';
+
+/* eslint-disable import/no-unresolved, import/extensions */
+import AggregatorContainer from 'terra-aggregator/lib/terra-dev-site/doc/common/AggregatorContainer';
+import AggregatorItem from 'terra-aggregator/lib/terra-dev-site/doc/common/AggregatorItem';
+import SimpleAggregatorItem from 'terra-aggregator/lib/terra-dev-site/doc/common/SimpleAggregatorItem';
+/* eslint-enable import/no-unresolved, import/extensions */
+
+const propTypes = {
+  size: PropTypes.string,
+};
+
+const items = Object.freeze([{
+  key: 'SECTION_0',
+  component: <AggregatorItem name="Slide Panel Section" disclosureType="panel" />,
+}, {
+  key: 'SECTION_1',
+  component: <AggregatorItem name="Modal Section" disclosureType="modal" />,
+}, {
+  key: 'SECTION_2',
+  component: <SimpleAggregatorItem name="No Disclosure Section" />,
+}]);
+
+const CombinedDisclosureExample = () => (
+  <div>
+    <ModalManager>
+      <SlidePanelManager>
+        <AggregatorContainer
+          items={items}
+        />
+      </SlidePanelManager>
+    </ModalManager>
+  </div>
+);
+
+CombinedDisclosureExample.propTypes = propTypes;
+
+export default CombinedDisclosureExample;