+++ packages/terra-app-delegate/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Moved examples to src
 
 2.2.0 - (May 22, 2018)
 ------------------