+++ packages/terra-slide-panel/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Moved examples to src
 
 2.7.0 - (May 22, 2018)
 ------------------